% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo_desempleo.R
\name{calculo_desempleo}
\alias{calculo_desempleo}
\title{Calculo desempleo}
\usage{
calculo_desempleo(encuesta)
}
\arguments{
\item{encuesta}{string con la ruta a un archivo SPSS}
}
\description{
Calculo desempleo
}
